apps/ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
redis==5.0.4
httpx==0.27.0
joblib==1.4.2
